+++ tasks/generate_docs.clj
@@ -1,12 +1,12 @@
 (ns generate-docs
   (:require
     [clojure.java.io :as io]
-    [clojure.set :as set]
     [clojure.string :as str]
     [doric.core :refer [table]]
     [noahtheduke.splint]
     [noahtheduke.splint.config :refer [default-config]]
-    [noahtheduke.splint.rules :refer [global-rules]]))
+    [noahtheduke.splint.rules :refer [global-rules]]
+    [noahtheduke.spat.pattern :refer [simple-type]]))
 
 (defn get-config [rule]
   (@default-config (:full-name rule)))
@@ -43,17 +43,17 @@
           lines (str/join \newline (cons (first lines) dedented-lines))
           [docs examples] (str/split lines #"Examples:")]
       (str (str/trim docs)
-           \newline \newline
-           "### Examples"
-           \newline \newline
            (when examples
-             (str "```clojure"
+             (str \newline \newline
+                  "### Examples"
+                  \newline \newline
+                  "```clojure"
                   \newline
                   (str/trim examples)
                   \newline
                   "```"))))))
 
-(defn render-configuration [rule]
+(defn build-styles [rule]
   (let [config (get-config rule)
         chosen-style (:chosen-style config)
         supported-style (:supported-styles config)]
@@ -61,17 +61,38 @@
                (not (and chosen-style supported-style)))
       (throw (ex-info "Need both chosen-style and supported-style" {:rule (:full-name rule)})))
     (when (and chosen-style supported-style)
-      (let [config {:name "`:chosen-style`"
-                    :default (str "`" chosen-style "`")
-                    :options (->> supported-style
-                                  (map #(str "`" % "`"))
-                                  (str/join ", "))}]
-        (str "### Configurable Attributes"
-             \newline \newline
-             (print-table [{:name :name :title "Name" :align :left}
-                           {:name :default :title "Default" :align :left}
-                           {:name :options :title "Options" :align :left}]
-                          [config]))))))
+      [{:name "`:chosen-style`"
+        :default (str "`" chosen-style "`")
+        :options (->> supported-style
+                      (map #(str "`" % "`"))
+                      (str/join ", "))}])))
+
+(defn build-other-configs [rule]
+  (let [config (get-config rule)
+        opts (-> config
+                 (dissoc :description :enabled
+                         :added :updated
+                         :guide-ref :link
+                         :chosen-style :supported-styles)
+                 (not-empty))]
+    (when opts
+      (mapv (fn [[k v]]
+              {:name (str "`" k "`")
+               :default (str "`" v "`")
+               :options (str/capitalize (name (simple-type v)))})
+            opts))))
+
+(defn render-configuration [rule]
+  (let [styles (build-styles rule)
+        other-configs (build-other-configs rule)
+        config (seq (concat styles other-configs))]
+    (when config
+      (str "### Configurable Attributes"
+           \newline \newline
+           (print-table [{:name :name :title "Name" :align :left}
+                         {:name :default :title "Default" :align :left}
+                         {:name :options :title "Options" :align :left}]
+                        config)))))
 
 (defn render-reference [rule]
   (let [config (get-config rule)